refactor(core): split System actions into StartedSystem for better state management
Refactored the `System` class by introducing a new `StartedSystem` class to encapsulate actions performed after system initialization. Updated the main program and tests to use `StartedSystem`. This improves code clarity and enforces proper state management.
