+++ exercise/C#/day23/ControlSystem.Tests/TestControlSystem.cs
@@ -25,38 +25,38 @@ public class TestControlSystem : IDisposable
     public void TestStart()
     {
         // The system has been started
-        _controlSystem.Action = SleighAction.Flying;
-        _controlSystem.Status = SleighEngineStatus.Off;
-        _controlSystem.StartSystem();
-        _controlSystem.Status.Should().Be(SleighEngineStatus.On);
+        var startedSystem = _controlSystem.StartSystem();
+        startedSystem.Action = SleighAction.Flying;
+        
         _output.ToString().Trim().Should().Be($"Starting the sleigh...{Environment.NewLine}System ready.");
     }
 
     [Fact]
     public void TestAscend()
     {
-        _controlSystem.StartSystem();
-        _controlSystem.Invoking(cs => cs.Ascend()).Should().NotThrow<ReindeersNeedRestException>();
-        _controlSystem.Action.Should().Be(SleighAction.Flying);
+        var startedSystem = _controlSystem.StartSystem();
+        startedSystem.Invoking(cs => cs.Ascend()).Should().NotThrow<ReindeersNeedRestException>();
+        startedSystem.Action.Should().Be(SleighAction.Flying);
         _output.ToString().Trim().Should().Be($"Starting the sleigh...{Environment.NewLine}System ready.{Environment.NewLine}Ascending...");
     }
 
     [Fact]
     public void TestDescend()
     {
-        _controlSystem.StartSystem();
-        _controlSystem.Ascend();
-        _controlSystem.Invoking(cs => cs.Descend()).Should().NotThrow<SleighNotStartedException>();
-        _controlSystem.Action.Should().Be(SleighAction.Hovering);
+        var startedSystem = _controlSystem.StartSystem();
+        startedSystem.Ascend();
+        startedSystem.Invoking(cs => cs.Descend()).Should().NotThrow<SleighNotStartedException>();
+        startedSystem.Action.Should().Be(SleighAction.Hovering);
         _output.ToString().Trim().Should().Be($"Starting the sleigh...{Environment.NewLine}System ready.{Environment.NewLine}Ascending...{Environment.NewLine}Descending...");
     }
 
     [Fact]
     public void TestPark()
     {
-        _controlSystem.StartSystem();
-        _controlSystem.Invoking(cs => cs.Park()).Should().NotThrow<SleighNotStartedException>();
-        _controlSystem.Action.Should().Be(SleighAction.Parked);
+        var startedSystem = _controlSystem.StartSystem();
+        
+        startedSystem.Invoking(cs => cs.Park()).Should().NotThrow<SleighNotStartedException>();
+        startedSystem.Action.Should().Be(SleighAction.Parked);
     }
 
     public void Dispose()
